requirements.txt
psutil~=5.9.8
art~=6.1
requests~=2.31.0
python-dotenv~=1.0.1
